college project

+++ src/main/java/com/noble/service/IStudentMgmntService.java
@@ -8,10 +8,10 @@ import com.noble.entity.Student;
 public interface IStudentMgmntService {
 
 	public String registerStudent(Student std); // to save Operations
-	public Map<Integer, String> getStudentdetails(); // to get student 
+	public Map<Long, String> getStudentdetails(); // to get student 
 	public List<Student> showStudents(); // to view students list
-	public Student getStudentById(Integer stdid); // to get Students by student Id
+	public Student getStudentById(Long stdid); // to get Students by student Id
 	public String editStudent(Student std); // editing Students
-	public String deleteStudent(Integer stdId); // deteting Student
+	public String deleteStudent(Long stdId); // deteting Student
 		
 }
